feat!: Removed legacy context manager
