--- .github/workflows/versioned-coverage.yml
@@ -1,38 +1,0 @@
-# This workflow is intended to be used to run versioned tests for different scenarios(i.e.- legacy context manager, etc)
-
-name: Nightly Versioned Scenario Runs
-
-on:
-  workflow_dispatch:
-  schedule:
-    - cron:  '0 9 * * 1-5'
-
-env:
-  # Enable versioned runner quiet mode to make CI output easier to read:
-  OUTPUT_MODE: quiet
-
-jobs:
-  legacy-context:
-    runs-on: ubuntu-latest
-
-    strategy:
-      fail-fast: false
-      matrix:
-        node-version: [18.x, 20.x, 22.x]
-
-    steps:
-    - uses: actions/checkout@v4
-    - name: Use Node.js ${{ matrix.node-version }}
-      uses: actions/setup-node@v4
-      with:
-        node-version: ${{ matrix.node-version }}
-    - name: Install Dependencies
-      run: npm install
-    - name: Run Docker Services
-      run: npm run services
-    - name: Run Legacy Context Versioned Tests
-      run: TEST_CHILD_TIMEOUT=600000 npm run versioned:legacy-context
-      env:
-        VERSIONED_MODE: --major
-        JOBS: 4 # 2 per CPU seems to be the sweet spot in GHA (July 2022)
-        SKIP_C8: true
