+++ lib/context-manager/create-context-manager.js
@@ -16,10 +16,6 @@ const logger = require('../logger')
  * the current configuration.
  */
 function createContextManager(config) {
-  if (config.feature_flag.legacy_context_manager) {
-    return createLegacyContextManager(config)
-  }
-
   return createAsyncLocalContextManager(config)
 }
 
@@ -28,13 +24,6 @@ function createAsyncLocalContextManager(config) {
 
   const AsyncLocalContextManager = require('./async-local-context-manager')
   return new AsyncLocalContextManager(config)
-}
-
-function createLegacyContextManager(config) {
-  logger.info('Using LegacyContextManager')
-
-  const LegacyContextManager = require('./legacy-context-manager')
-  return new LegacyContextManager(config)
 }
 
 module.exports = createContextManager
