+++ lib/feature_flags.js
@@ -15,7 +15,6 @@ exports.prerelease = {
   reverse_naming_rules: false,
   undici_async_tracking: true,
   unresolved_promise_cleanup: true,
-  legacy_context_manager: false,
   kafkajs_instrumentation: false
 }
 
@@ -43,4 +42,4 @@ exports.released = [
 ]
 
 // flags that are no longer used for unreleased features
-exports.unreleased = ['unreleased']
+exports.unreleased = ['unreleased', 'legacy_context_manager']
