+++ lib/instrumentation/core/globals.js
@@ -5,7 +5,6 @@
 
 'use strict'
 
-const asyncHooks = require('./async-hooks')
 const symbols = require('../../symbols')
 
 module.exports = initialize
@@ -63,8 +62,4 @@ function initialize(agent, nodule, name, shim) {
       return original.apply(this, arguments)
     }
   }
-
-  // This will initialize the most optimal native-promise instrumentation that
-  // we have available.
-  asyncHooks(agent, shim)
 }
