+++ lib/instrumentation/core/timers.js
@@ -11,48 +11,8 @@ const Timers = require('timers')
 
 module.exports = initialize
 
-function initialize(agent, timers, _moduleName, shim) {
-  const isLegacyContext = agent.config.feature_flag.legacy_context_manager
-
-  if (isLegacyContext) {
-    instrumentProcessMethods(shim, process)
-    instrumentSetImmediate(shim, [timers, global])
-  }
-
+function initialize(_agent, timers, _moduleName, shim) {
   instrumentTimerMethods(shim, [timers, global])
-}
-
-/**
- * Sets up instrumentation for setImmediate on both timers and global.
- *
- * We do not want to create segments for setImmediate calls,
- * as the object allocation may incur too much overhead in some situations
- *
- * @param {Shim} shim instance of shim
- * @param {Array<Timers,global>} pkgs array with references to timers and global
- */
-function instrumentSetImmediate(shim, pkgs) {
-  pkgs.forEach((nodule) => {
-    if (shim.isWrapped(nodule.setImmediate)) {
-      return
-    }
-
-    shim.wrap(nodule, 'setImmediate', function wrapSetImmediate(shim, fn) {
-      return function wrappedSetImmediate() {
-        const segment = shim.getActiveSegment()
-        if (!segment) {
-          return fn.apply(this, arguments)
-        }
-
-        const args = shim.argsToArray.apply(shim, arguments, segment)
-        shim.bindSegment(args, shim.FIRST)
-
-        return fn.apply(this, args)
-      }
-    })
-
-    copySymbols(shim, nodule, 'setImmediate')
-  })
 }
 
 /**
@@ -105,37 +65,6 @@ function wrapClearTimeout(_shim, fn) {
  */
 function recordAsynchronizers(shim, _fn, name) {
   return new RecorderSpec({ name: 'timers.' + name, callback: shim.FIRST })
-}
-
-/**
- * Instruments core process methods: nextTick, _nextDomainTick, _tickDomainCallback
- * Note: This does not get registered when the context manager is async local
- *
- * @param {Shim} shim instance of shim
- * @param {process} process global process object
- */
-function instrumentProcessMethods(shim, process) {
-  const processMethods = ['nextTick', '_nextDomainTick', '_tickDomainCallback']
-
-  shim.wrap(process, processMethods, function wrapProcess(shim, fn) {
-    return function wrappedProcess() {
-      const segment = shim.getActiveSegment()
-      if (!segment) {
-        return fn.apply(this, arguments)
-      }
-
-      // Manual copy because helper methods add significant overhead in some usages
-      const len = arguments.length
-      const args = new Array(len)
-      for (let i = 0; i < len; ++i) {
-        args[i] = arguments[i]
-      }
-
-      shim.bindSegment(args, shim.FIRST, segment)
-
-      return fn.apply(this, args)
-    }
-  })
 }
 
 /**
