+++ test/unit/context-manager/create-context-manager.test.js
@@ -8,7 +8,6 @@
 const { test } = require('tap')
 
 const createImplementation = require('../../../lib/context-manager/create-context-manager')
-const LegacyContextManager = require('../../../lib/context-manager/legacy-context-manager')
 const AsyncLocalContextManager = require('../../../lib/context-manager/async-local-context-manager')
 
 test('Should return AsyncLocalContextManager by default', (t) => {
@@ -18,15 +17,5 @@ test('Should return AsyncLocalContextManager by default', (t) => {
   })
 
   t.ok(contextManager instanceof AsyncLocalContextManager)
-  t.end()
-})
-
-test('Should return LegacyContextManager when enabled', (t) => {
-  const contextManager = createImplementation({
-    logging: {},
-    feature_flag: { legacy_context_manager: true }
-  })
-
-  t.ok(contextManager instanceof LegacyContextManager)
   t.end()
 })
